lambda/requirements.txt
fastapi
uvicorn[standard]
requests
yt-dlp
google-auth
google-api-python-client